# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "wallet",
        "title": "wallet",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/api/v1",
    "paths": {
        "/user/login": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "parameter",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginOutput"
                        }
                    }
                }
            }
        },
        "/wallet/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the user's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get the user's balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.BalanceOutput"
                        }
                    }
                }
            }
        },
        "/wallet/deposit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User deposits",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "User deposits",
                "parameters": [
                    {
                        "description": "parameter",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.DepositReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/wallet/record": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user transaction records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get user transaction records",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.RecordOutput"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/transfer": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User Transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "User Transfer",
                "parameters": [
                    {
                        "description": "parameter",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.TransferReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/wallet/withdraw": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User withdraw",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "User withdraw",
                "parameters": [
                    {
                        "description": "parameter",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.WithdrawReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "types.BalanceOutput": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "types.DepositReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount to deposit",
                    "type": "number"
                }
            }
        },
        "types.RecordOutput": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "amount",
                    "type": "number"
                },
                "created_at": {
                    "description": "Creation time",
                    "type": "string"
                },
                "describe": {
                    "description": "description",
                    "type": "string"
                },
                "id": {
                    "description": "wallet record id",
                    "type": "integer"
                },
                "transaction_type": {
                    "description": "transaction type 0=unknown 1=income 2=expend",
                    "type": "integer"
                }
            }
        },
        "types.TransferReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount to transfer",
                    "type": "number"
                },
                "target_user_id": {
                    "description": "Transfer target user ID",
                    "type": "integer"
                }
            }
        },
        "types.UserInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "types.UserLoginOutput": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/types.UserInfo"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "types.UserLoginReq": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "login email",
                    "type": "string"
                },
                "password": {
                    "description": "login password",
                    "type": "string"
                }
            }
        },
        "types.WithdrawReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount to withdraw",
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Please enter your Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}